move connection_options to its own util
